+++ Sources/SVGEngine.mm
@@ -539,9 +539,22 @@ CF_RETURNS_RETAINED CGMutablePathRef pathDefinitionParser::parse()
         if([cmdBuf length] > 1) {
             scanner.scanLocation -= [cmdBuf length]-1;
         } else {
-            for(float operand;
-                [scanner scanFloat:&operand];
-                _operands.push_back(operand));
+            while (!scanner.isAtEnd) {
+                NSUInteger zeros = 0;
+                while ([scanner scanString:@"0" intoString:NULL]) { ++zeros; }
+                // Start of a 0.x ?
+                if (zeros > 0 && [scanner scanString:@"." intoString:NULL]) {
+                    --zeros;
+                    scanner.scanLocation -= 2;
+                }
+                for (NSUInteger i = 0; i < zeros; ++i) { _operands.push_back(0.0); }
+
+                float operand;
+                if (![scanner scanFloat:&operand]) {
+                    break;
+                }
+                _operands.push_back(operand);
+            }
         }
 
 #ifdef SVG_PATH_SERIALIZER_DEBUG